requirements.txt
pytz
openai-whisper
